prelude: Refactor javacd_jar_creator to prepare for write_json changes
Summary:
We have an upcoming change where `write_json` run with `with_inputs` will attach all bound inputs to the resulting artifact. This means any unbound artifacts at this point are not going to be included, so we need to add those manually to the following run action. This includes all `OutputPaths` artifacts and in some cases `path_to_class_hashes`.

This change is compatible with the existing buck2 version and the future `write_json` changes.

Reviewed By: ndmitchell

Differential Revision: D39970836

fbshipit-source-id: 7a49e7eedfc66bcd066aeccbadec3cce7cbd1fa1