use menhir unconditionally
Summary:
D40022938 makes menhir available to the ocaml toolchain. this diff replaces use of `ocamlyacc` with it.

i had to do a bit of work on the flow semver grammar rules to get the tests passing with menhir. i think the resulting grammar is ever so slightly more correct (& verified against ocamlyacc)  so shouldn't be contentious

Reviewed By: samwgoldman

Differential Revision: D40025968

fbshipit-source-id: bfbbbeacaacfa4fd359432145ec489488ed4f4b0