Migrate genrule to use `with_inputs=True`
Summary:
This is part of a few diffs I'm putting up to migrate a few instances of `ctx.actions.write` to use the new `with_inputs` flag, when appropriate. This will attach any bound artifacts referenced in the write action to be attached to the resulting artifact as an `associated artifact`. This removes the need to add such inputs to the following cmd args entity with `hidden` as they will be accessible through the artifact.

This diff migrates the genrule implementation.

Reviewed By: krallin

Differential Revision: D39977200

fbshipit-source-id: 5d45977556f02e73c523a270d95e40ff2e186b5e