Expand the supported type on label_matches_build_target_pattern
Summary: To make it easier to utilize this method in BXL, expand the supported type to include `target_label`, which we need to compare to build target patterns in project generation. Both `label` and `target_label` support the same API as far as the methods called on it in the function are concerned.

Reviewed By: ndmitchell

Differential Revision: D40077001

fbshipit-source-id: 5e7918d776b1041a4a27b956afc27e4d6f4e1de0