Specify that we are only going to build native code for non-primary platform
Summary:
When building Android APKs, if we are doing a multi-cpu build (i.e. we have more than one `cpu_filter`) then we do a split-transition on the `deps` of the top-level binary. This allows us to build the native libraries with all the specified cpus, but it also means that throw away all of the analysis of Java-based code for all the resulting cpu platforms except for the first one.

To reduce this "wasted" analysis, we can have a flag that specifies that we are going to throw away the Java-based part of the analysis, and not do most of it (we just do the part that is relevant to native libraries).

This gets a little messy, mostly because of `android_instrumentation_apk`, which has an attribute "apk" which is the "APK-under-test". We want the `android_instrumentation_apk` to have the same `cpu_filters` as the `apk`, but I don't know how to access the `cpu_filters` from the `apk` before we need to use them in our split transition.

Tried:
1) Macro layer. This doesn't work because at that point the `apk` attr is just a string (that represents a build target).
2) By accessing the `apk` attribute in the split transition itself. This doesn't work as the `apk` attribute can't be coerced at that point, it fails with: `Attribute cannot be converted to Starlark value: "fbsource//fbandroid/buck2/tests/good/instrumentation_test:apk_under_test"`.

By the time we can read the `cpu_filters` attribute from the `apk`, the split transition has already occurred.

That means that for `android_instrumentation_apk`, we don't know which was the primary platform for the `apk`, and which other platforms it uses. Therefore, we need to do analysis of the `apk` without specifying that secondary platforms are only going to build native code, and then for the `android_instrumentation_apk` we specify all possible platforms.

Reviewed By: ianlevesque

Differential Revision: D37550647

fbshipit-source-id: 942c4dca7da91a23b73db0a9dcc9f3518f087cb0