Only do class ABI generation for Kotlin targets unless specified
Summary:
In buck1, we don't create a source-only ABI unless it's explicitly set on a target with language = "KOTLIN".

In buck2, if a target is set with language = "KOTLIN", but it actually only has Java sources, then right now we do try to create a source-only ABI. That can lead to issues where the source-only ABI fails to compile because we haven't told it the correct source-only ABI deps to use.

Instead, just use the same behavior as buck1, and don't create one at all.

Reviewed By: navidqar

Differential Revision: D42098016

fbshipit-source-id: 6d09aae5712a9cc0af83a1bfaa6ec2d43365f2ea