use absolute path for deployment scripts
Summary: This is a workaround to enable Buck2 on our ARVR deployment scripts. I attempted in D41468559 to rewrite the rules to use the recommended approach of staying away from genrules with intermediate commands but it only works for individual deploy targets, not for target collections. In the meantime this is the quickest way to unblock us.

Differential Revision: D42207765

fbshipit-source-id: 3b4d72526cb85745428c872666e1c6f41c832151