allow dSYM output path to be customizable
Summary:
What I'm trying to achieve is that dSYM for `apple_test` target is named in sync with XCTest bundle (i.e. it is "foo.xctest.dSYM" for "foo.xctest" bundle). That output is similar to what v1 has and also is needed for fdb/idb to automatically find dSYM when `--install-dsym-test-bundle` argument is passed without any value.

No behavior change (see it is used in next diff).

Reviewed By: milend

Differential Revision: D42344577

fbshipit-source-id: 8597f6ef1892968b89d452fc155b0396651be3d0