Add wrapper bat file for hermetic msvc compilation
Summary: cl.exe's equivalents of clang's `-fdebug-prefix-map` and related flags require absolute paths. It is not sufficient to just pass '.'. Since buck2 can't generate absolute paths at analysis time, we need a wrapper to compute the paths at execution time. A simple bat file is sufficient, so we set it up as a target and pass it through the cxx_toolchain as a default argument. Wrap the compiler with it if we're using a vanilla windows compiler and not a clang variant.

Reviewed By: KapJI

Differential Revision: D42510765

fbshipit-source-id: 2c5b0e1ec9eb89515bdaa9a09d9b0d1f9ea4e1e3