Fix cross platform build logic for command_alias
Summary:
Follow up on D41370266 (https://github.com/facebookincubator/buck2-prelude/commit/37752b6ec36a68c169053cb3f7ba359b677a22b6)

command_alias needs to produce a script that is runnable on the exec platform, but it needs to run shell commands (in the target platform) to produce the script. This means, if we create a batch script on a unix machine, we need to run sed in the process. If we create a shell script on a windows machine, we need to run a sed-like batch script.

krallin had a change to align target/exec platforms when possible (D41182948, P604218519), but for some reason this does not work. While this is still being investigated, we actually are hitting this problem with D42589792 where we try to build for Windows on Linux. This diff resolves the command_alias problem, though we still need to understand why the biasing doesn't work.

Reviewed By: krallin

Differential Revision: D42691438

fbshipit-source-id: 90e122f90e615f3f8f27e2dada7d92e760f8544d