Don't pass False as the default to read_config
Summary: I don't think read_config should allow you to pass a False, since config is either a String or returns None for not found. If you are then going to boolean on the result, using None as the default makes much more sense, since then you are only dealing with None and valid values. Therefore, if we were passing False as the default, and no default (implicitly None) worked just as well, do that.

Reviewed By: IanChilds

Differential Revision: D42707940

fbshipit-source-id: a2a7d06edda654bd00f65ae6f63466d0173f609c