No doc tests during cross build
Summary: I tried `cargo test --doc --target aarch64-unknown-linux-gnu` and Cargo silently did not run doc tests. We could probably make this work, but it seems low value, and Cargo not running them tells me we'd be very likely to hit issues with this not being supported well in rustdoc.

Reviewed By: zertosh

Differential Revision: D42609594

fbshipit-source-id: 99248ff5aa9c6bc56f56bb7e5b278985f5501cd6