java|kotlin|robolectric tests shouldn't do anything if _build_only_native_code is True
Summary:
It sort of surprised me that this is needed, but when we do `testsof` on a configured dep, the tests are configured with the same configuration.

I think this does make sense, since it means that the test deps will get the same config as the binary/library deps, and so we don't have two separate graphs of all the deps.

However, it does mean that we have to account for `_build_only_native_code`, which for test rules just means that we shouldn't return any providers (except `DefaultInfo` which we have to produce.

cc cjhopman in case I got anything wrong here.

Reviewed By: jselbo

Differential Revision: D42740475

fbshipit-source-id: d1f23ad7dcc52964be3b22bfb01cdb45ab75ef46