Return deps as a dict when buck2.android_force_single_cpu is enabled
Summary:
Previously, if this flag was turned on then we would return just the configuration: https://www.internalfb.com/code/fbsource/[d64eca33e12f]/fbcode/buck2/prelude/android/configuration.bzl?lines=103

This meant that analysis would fail as it expects `deps` to be a dict of `{platform_name: configuration}`.

Instead, we return a dict with just one `cpu` inside of it.

Reviewed By: krallin

Differential Revision: D42775741

fbshipit-source-id: ddb993d4eab02a19c198507436a91bd1f22b343a