add swift object files of shared output to debug info
Summary:
The cxx rules currently propagate linked object files to `external_debug_info` for Darwin only for archive rules:

https://www.internalfb.com/code/fbsource/[65d4ca8594b6]/fbcode/buck2/prelude/cxx/cxx_library.bzl?lines=924-928

When building the `[shared]` output of an `apple_library` we will not have any archives, so the debug info is missing. Explicitly adding it to the cxx rule params ensures it is propagated through correctly, regardless of archive use.

Reviewed By: milend

Differential Revision: D42884638

fbshipit-source-id: 110b4e8455e352e3f75a2e214da16e1bd71dff9e