Avoid unnecessary runtime deps on dlopen-able libs
Summary:
Previously, we'd put all dlopen-enabled libs onto the link line of the main
executable, even if it doesn't actually depend on it (e.g. a dlopen-enabled
lib is often just referenced by a `python_library` and so has no native
dependents that it'd need to link with).

This diff fixes by keeping track of the two types of search roots being
passing to `get_filtered_labels_to_links_map`:
1) deps that should always be included in the link processing and
2) roots that should be searched for the first occurrence of a linkable
    node which is part of the relevant link group and which should be
    treated as a "dep", as per (1).

Reviewed By: christycylee

Differential Revision: D42854938

fbshipit-source-id: 5c814e4fc4401f0b863400be9026a523e10c8a0c