Fix link group mapping computation with dlopen-libs/extensions
Summary:
Make sure we include dlopen-enabled libs and unembeddable extensions
in the linkable graph we use to compute link group mappings.  This fixes a
bug where we'd incorrectly determine a node didn't belong to a link group
(which it should belong to) if it was only reachable via a dlopen-enabled
library or unembeddable extension.

Differential Revision: D42927070

fbshipit-source-id: 5f8adfa9eb8d6a9de40ab4c38203da47a74b72bd