MergeAndroidResourcesExecutableMain returns a .src.zip of the sources
Summary:
The general way to pass multiple sources to Java compilation is to use a `.src.zip` file containing all of the sources. By doing this, we are able to avoid having do a dynamic action that reads the `src_listing` that we created and then copies the outputs to somewhere where `buck2` can keep track of them.

This gets rid of about 5900 commands in the `fb4a` build (out of 312k, so slightly less than 2%).

Reviewed By: navidqar

Differential Revision: D42963111

fbshipit-source-id: 6fa34639b44a74cef2096b4da988becb4b8753ea