Expose all package styles as subtargets
Summary:
This will let people force a particular `package_style` without having to pass in a global `-c python.package_style=foo` option which invalidates a bunch of the graph and does result in re-analysing.

Some targets don't build with certain package_styles. For these, the subtargets are still exposed, but trying to build them will cause a build-time error. If the main target doesn't build with the default package style though, the error will be emitted at analysis time.

#pyredisabledependentjobs

Reviewed By: krallin

Differential Revision: D41999601

fbshipit-source-id: 820b9277c3e0544cfa51ec51e01e88b00fc83867