Generate manifest modules in prelude
Summary:
Generation of the `__manifest__` module is done in two places:
- added as a synthetic source dependency in fbcode macros
- as part of make_par

This diff adds a third place, with the intent to consolidate the above two. Build info is passed down from fbcode macros as an attr, which gets passed into a script that makes `__manifest__.py`.

A followup diff will remove the synthetic manifest dependency, and then we can look at removing this functionality from make_par.

#pyredisabledependentjobs

Reviewed By: krallin

Differential Revision: D42503627

fbshipit-source-id: 9973ecb7cb62297e6798af7fa48a4d98bd576ee1