Don't propagate unused preferred-linkage mappings
Summary:
When a preferred-linkage value is set in a mapping, it's used to force
a library and it's transitive deps to be evicted from the binary.  This diff
fixes a bug where we'd do this transitive eviction even if the root node
that the mapping refers to isn't in the deps of the top-level binary.

https://fb.workplace.com/groups/linker.support/posts/931075048059338/

Differential Revision: D43026308

fbshipit-source-id: 440350f8e13396d612626f0a6d2569cc6cd8b704