Add starlark rules for Julia
Summary:
This is a series of Starlark rules (and a corresponding toolchain) for the Julia language.

The toolchain is currently modeled after the `Go` toolchain, although future efforts should expand it to match the `rust` toolchain for maximum compatibility (not needed for now). #Thanks to krallin and diliop for helping with debugging various aspects of the Julia toolchain and rule definitions!

The current rules that are supported are `julia_binary`, `julia_library`, and `julia_test`.

The `julia_binary` rule is rather robust, and even supports passing various julia flags (e.g. `-O3`). The other rules are not fully implemented. But their current state serves as a good starting point for future iterations.

Reviewed By: krallin

Differential Revision: D42822046

fbshipit-source-id: 10350d2b3c4670daa2254c1007065c1da7435b67