migrate any()/all() functions with inlined ListCmp's in prelude to is_any/is_all
Summary:
In the next diff, we are going to enable our new starlark linter to warn against `any()`, `all()` functions using inlined list comprehensions.

For scenarios like this:

```
_ = any([i % 2 == 0 for i in range(100000)])
```

The list comprehension would lead to a new list of 100000 booleans, and would only end-up checking 1. Replacing it with:

```
_ = is_any(lambda i: i % 2 == 0, range(100000))
```

would lead to 0 new allocations.

Reviewed By: IanChilds

Differential Revision: D43146708

fbshipit-source-id: 176f3ab660ac7b26c869b6c049cdb8d45bec7dc4