Prioritize generating results before artifacts.
Summary:
This change is not "the right one", as what we should aim for is that those artifact / logs generation be handled by whatsapp specific providers.

It's not my priority to do this at the moment, but this change is a small win nonetheless.

What we do here is change the order of execution between generating the result file / uploading -preparing the artifacts.

Reviewed By: TheGeorge

Differential Revision: D43234237

fbshipit-source-id: 56544f080d7f5f2f13b119dc6f149aa597a444bb