Put weight_estimate in same file as filtered_class_names_file
Summary:
Not totally sure about this one, but the idea is that by writing the weight estimate to the `filtered_class_names_file`, we have to materialize fewer artifacts in order to decide which dex to assign each `.class` to.

In particular, right now for `fb4a` we have about 32k pre-dexed libraries. For each of these, we need to read both the `weight_estimate` and the `filtered_class_names`. By putting those in the same file, we only have to materialize 32k files, instead of the 64k that we were materializing before.

On a fresh build after `buck2 clean`, the materialization is pretty slow (>10s), so this might have a reasonable impact. krallin does have some changes that might speed this up, but this probably doesn't do any harm.

Reviewed By: navidqar

Differential Revision: D43272865

fbshipit-source-id: b449088bfe5ffa6aa1ec4377ffb6fd293f08f53d