cxx_library should return an empty SharedLibraryInfo when cxx_platform_supported is False
Summary:
In `java_library`, when checking our dependencies are valid we use the existence of a `SharedLibraryInfo` to mean that a dependency is a native library (and therefore valid): https://www.internalfb.com/code/fbsource/[3788d901f8bf]/fbcode/buck2/prelude/java/java_library.bzl?lines=436-437

If we're building on a platform where a `cxx_library` is not supported, it doesn't mean that we should mark the dep as invalid, but that is what happens right now, e.g. P623743800 - we're building for a robolectric test using the `android-host-test` platform, so `cxx_supported_platforms` doesn't match https://www.internalfb.com/code/fbsource/[63637f9d1f53]/fbandroid/java/com/facebook/common/fury/jni/BUCK?lines=61 and we complain that the dep isn't a native library dependency.

To fix that, it seems reasonable to just produce an empty `SharedLibraryInfo` when the platform is not supported.

Reviewed By: navidqar

Differential Revision: D43307051

fbshipit-source-id: cfa9ad89215831f0d6ab29f237b4e575dde6c23c