JavaCD and KotlinCD support for per-class dep-files
Summary:
This adds a new dep-file option `per-class` to JavaCD and KotlinCD.

When enabled, for any ABI on the classpath, if that ABI has a corresponding ABI directory representation, we add the ABI directory representation as an input.

We pass a mapping of ABI-to-ABI-directory-representation our dep-file wrapper `used_classes_to_dep_file.py`, and for any used `.jar`s, the wrapper  will look up whether the `.jar` has a directory representation. If it does not, the wrapper will just make the `.jar` file used as before. If it does, then the wrapper will take the list of used `.class` files and mark those as used in the directory representation.

Reviewed By: christolliday

Differential Revision: D43118078

fbshipit-source-id: 0cab2cf47883ad3f840ba8cc4b4c59882534b403