e2e: fix test_js test
Summary:
Some of the targets are not building due to the output type not being correct. (we expect directory but it is not)

Introduced in D43328609 (https://github.com/facebookincubator/buck2-prelude/commit/625847490ce432801b3fe1aa7de408db05e9cee4), but since some of the dir=true markings are correct, fixing it forward.

Reviewed By: blackm00n

Differential Revision: D43398283

fbshipit-source-id: ec89c7dcfc4be586990f3faf9e2606a3fa166ce1