Move "buck test causes doctests to run" from fbcode_macros into prelude
Summary:
The reason this went in fbcode_macros to begin with was because we started with applying doc tests to *only* `common/rust/` in D42516932, before later expanding that to all targets in D42565392.

This diff makes `buck2 test ...` include doctests in repos which do not use fbcode_macros, such as https://github.com/dtolnay/cxx. For fbcode there should be no change in behavior.

This is a prerequisite for 2 desirable improvements I have on the way:

1. Move unit test setup from fbcode_macros into buck2 prelude. Currently OSS use of buck1 does not get unit testing on their `rust_library` and `rust_binary` targets. I am moving fbcode_macros' `buck2 test :crate-unittest` to a native `buck2 test :crate[unittest]`.

2. Following that, I want to make `buck2 test :crate[doc]` run *only* doctests and `buck2 test :crate[unittest]` run *only* unit tests. `buck2 test :crate` will continue to run both. Currently all of these commands cause both test kinds to run so there is no way to test only one or the other, a regression compared to Cargo's `cargo test --lib` and `cargo test --doc`.

Reviewed By: diliop

Differential Revision: D43336765

fbshipit-source-id: 4ea44aef1e33ff51e860de72188cc55014d00228