Add an attribute for setting environment during doc tests
Summary:
This works like the existing `test_env` (which is set during compilation and execution of unittests) and `env` (which is set for the ordinary library build in addition to unit tests and docs).

```
rust_library(
    ...,
    doc_env = {
        "FOO": "$(location :foo)",
    },
)
```

I wanted to use this in the doc tests of `codegen_includer_proc_macro` which expands based on `$OUT_DIR`. I didn't want to use ordinary `env` because that would make bugs like replacing `env::var("OUT_DIR")` (OUT_DIR during the build of the macro caller) with `env!("OUT_DIR")` (OUT_DIR during build of the proc macro dylib).

https://www.internalfb.com/code/fbsource/[2e8ecd86d9327814be8dafebc8402ff7b4139d54]/fbcode/common/rust/shed/codegen_includer_proc_macro/src/lib.rs?lines=22-23

https://www.internalfb.com/code/fbsource/[2e8ecd86d9327814be8dafebc8402ff7b4139d54]/fbcode/common/rust/shed/codegen_includer_proc_macro/src/lib.rs?lines=36-40

The `doc_env` attribute will be covered by the existing check that disallows it from being present when `doctests = False` is set.

https://www.internalfb.com/code/fbsource/[fea5dc8323f0bb561f746db029def36fe3588c30]/tools/build_defs/fbcode_macros/build_defs/lib/rust_common.bzl?lines=272-276%2C687-699

Reviewed By: zertosh

Differential Revision: D43511518

fbshipit-source-id: ad32227009d562036a199c6c359b14250a4c91d9