Make it so that the android APK 'classpath' contains any Java library with a .jar
Summary:
The `classpath` should consist of any `JavaPackagingDep` that has a `.jar` associated with it (i.e. any "java" dep that produces an output).

Previously, `android_apk` was filtering out anything that didn't have a `.dex`, since that's all that we actually need to package.

Instead, just filter out anything that doesn't have a `.dex` for the "dex" steps, and keep all the java packaging deps for everything else.

Reviewed By: scottcao

Differential Revision: D43940725

fbshipit-source-id: 29f3dab8307ac61772bb0b77dba06462318aa8c3