buck2: prelude: explicitly close outputs in Python scripts
Summary:
Like it says in the title. If your disk is full, you don't necessarily get
errors on `write`, you may get them on `close`, and if you ignore errors on
`close` (which the [destructor does in Python][1]), then you'll silently ignore
the fact that your data was not written.

[1]: https://github.com/python/cpython/blob/a703f743dbf2675948e59c44fa9d7112f7825100/Lib/_pyio.py#L419-L430

Reviewed By: ndmitchell, milend

Differential Revision: D44054051

fbshipit-source-id: f985a7010b5b613e27d675d4e9fc963c8452fbd3