Make sure 'includes' and 'libraryjars' are present in proguard output
Summary: This is a bit of a hack, but Redex will read the proguard command line and use the artifacts that it finds in it. Therefore we need to ensure that the things it finds are materialized, so create symlinks to those things. For now, those things are the 'includes' and 'libraryjars'.

Reviewed By: agampe

Differential Revision: D44027206

fbshipit-source-id: dd8b871e7061d245ecdb72c3ec7661f2f5360bfc