Use exec in command alias trampoline
Summary:
It's the last command being run by the script, so we may as well `exec`
it and save an unnecessary process creation. I wasn't expecting it to
make any real difference in practice, but I A/B tested it just for fun
on my Mac (where process spawning is slower, with the change in
P649047694 to establish a marker. The A warmup command bumps the cache
invalidation counter and restores the last line to its original form,
and the B warmup command bumps the cache invalidation counter and
prefixes the last trampoline line with an `exec`.)

```
absh \
  -a 'buck2 build --local-only "//fbandroid/native/fb:fb[shared]"' \
  -b 'buck2 build --local-only "//fbandroid/native/fb:fb[shared]"' \
  -A "perl -pi -e 's/(?<=cache invalidate )(\\d+)/(\$1+1)/e;s/\\x27.*\\x22\\\$\\{ARGS/\\x27\\x22\\\$\\{ARGS/' /Users/smeenai/fbsource/fbcode/buck2/prelude/command_alias.bzl" \
  -B "perl -pi -e 's/(?<=cache invalidate )(\\d+)/(\$1+1)/e;s/\\x27.*\\x22\\\$\\{ARGS/\\x27exec \\x22\\\$\\{ARGS/' /Users/smeenai/fbsource/fbcode/buck2/prelude/command_alias.bzl"

Time (in seconds):
A: n=6 mean=21.229 std=0.871 se=0.389 min=20.118 max=22.355 med=21.005
B: n=6 mean=20.718 std=0.588 se=0.263 min=20.034 max=21.434 med=20.791
A: distr=[  ▌               ▌▌          ▌                      ▐    ▐]
B: distr=[█             ▌         ▌    ▐     ▐                       ]
B/A: 0.976 0.921..1.034 (95% conf)
```

So no significant difference, but still worth not doing unnecessary work IMO.

Reviewed By: IanChilds

Differential Revision: D44001505

fbshipit-source-id: ae40792e84c2fe3f6689bd08765e6b8020128b48