Enable cuda in buck2 builds.
Summary:
We can't directly rely on python_binary as a member of
a cxx_toolchain, because python_binary depends on cxx_toolchain itself.
There's a helper called python_bootstrap_binary which was added for
this case which does *not* depend on cxx_toolchain and breaks the cycle.
Port arvr's cuda wrapper to python_bootstrap_binary so we can use cuda
in our buck2 builds.

Fix the misc buck2 deficiencies that block this from working too.

Reviewed By: akrieger

Differential Revision: D43522656

fbshipit-source-id: 07825731f1f6d539fc8fac9fc608ba4c085f6ff9