Rework cmd_args so it supports incoming SwiftSupport dir.
Summary:
Currently, `apple_package` creates an IPA by declaring an output for `Payload`, going to the `.parent()` and zipping from there.

To support `SwiftSupport` folder, we are going to need a cleaner way to zip these togethers. As such, I've created: `_get_ipa_contents` which returns an unzipped version of the directory that we can easily pass to the `zip` command to fixup.

Reviewed By: chatura-atapattu

Differential Revision: D44520687

fbshipit-source-id: 7764f21b210f554bf0e1c171e251cf3cbe9ba1e0