Add Symbols/ dir to packages contain watch Executable.
Summary:
When packaging a `.app` containing a Watch Extension, the `.ipa` is expected to have an empty `Symbols/` directory at the top-level. This adds the functionality to buck2.

IMO this is a little more accurate than the buck1 implementation because it actually inspects the assembled bundle to determine if the folder is needed (as opposed to string-matching the platform name).

Reviewed By: milend

Differential Revision: D44675633

fbshipit-source-id: c7542a0cec7368a4ad3bebdcc02746f90e954586