Make coverage collection work for Buck2 + Robolectric
Summary:
Updated Buck macros to generate class name to source mappings for Robolectric tests.

After doing that, coverage was failing because Jacoco was trying to process JARs which didn't exist locally. Fixed by ensuring that mappings are only generated for JARs which are present at runtime -- we shouldn't generate mappings for "provided" or "provided_exported" deps.

Reviewed By: IanChilds

Differential Revision: D44843318

fbshipit-source-id: b8a3fa862c3c9561e19a1d5b5c682ef60a5e2741