Define a scrubber function and provide it
Summary:
We no longer want to provide `apple_focused_debugging` as a `link_postprocess` attribute. This is because post processing the executable changes its output, and thus triggers relinking up the link group tree.

Instead, we make focused debugging a first party integration on `apple_bundle`. And when provided as an `focused_debugging` attribute on `apple_bundle, will be used to scrub the binary and filter debug data (next diff).

To facilitate this, we
1. elevate the definition of the `apple_focused_debugging` rule. It must provide a `RunInfo`
2. in `AppleFocusedDebuggingInfo`, set a `scrub_binary` function, which when provided a ctx and executable, will scrub the binary.

Reviewed By: blackm00n

Differential Revision: D45130220

fbshipit-source-id: 3e18396194656f61717bc7d2ead7f09eb1004238