switch to static-pic link style
Summary:
to avoid `relocation R_X86_64_32 cannot be used against local symbol; recompile with -fPIC` style errors (in circleci on unbuntu) when embedding rust libraries in ocaml binaries switch the link-style to `static-pic` in `ocaml.bzl` everywhere.

re-enable the targets that were broken due to these errors.
- ocamlrep
- buck2 (extend only, embed so far unchanged)

the buck2 target re-enablements (embed directories) were reviewed in D45292084. i've decided to fold that changeset into this one.

i have also put `-cclib -lpthread` on ocaml link lines in the `ocaml_binary` rule. this is because only in sandcastle linux oss without it there are build errors for missing symbols and doing this overcomes it. i have not been able to reproduce the failure elsewhere. since it doesn't break other builds and fixes that failure it's all i can think of to do for now.

Reviewed By: themarwhal

Differential Revision: D45291308

fbshipit-source-id: 32c6acf9066ce65375b9d43e2445ca51f5c3650f