offline builds: make pxl_library not depend on hg at build time
Summary:
One of the blocker I'm running into with building `fbcode//manifold/blobstore:blobstore_svc` in a fully-offline environment is that one of its dependencies brings in `pxl_library` which currently requires invoking hg at build time.

As near as I can tell, the only usage of this is to run `hg root` to discover the repo root path so it can construct an include path for the `pxl` tool.

Instead, let's allow this to run from the project root in buck2's genrule.

The correct approach here would be to explicitly spell out which directories `pxl_library`s need and pass that along to the `pxl` tool. Alternatively, this could be written as a native buck2 rule which allows for the same thing.

Separately, this genrule still can't run on RE because it relies on a dotslash version of `pxl` which has everstore as a backend, which can't run on RE.

Differential Revision: D45238388

fbshipit-source-id: c0dd503f47985e68aecdcb7412982c4c0edcb2f8