Make Swift compatible with Link Groups
Summary:
**Problem**: The swift integration currently passes the swiftmodule linker args as extra_linker_args to the `cxx_library` implementation. These args must include any swift includes of transitive libraries.

Thus, when we encounter an `apple_library` that is linked as a link group shared library, this includes all swiftmodules that are transitive deps. Link groups however, should only ever include deps that are being linked (which may not be all deps).

See internal example below.

When the interface of a swift class is modified, this generates a different swiftmodule file, which as an input to the apple_library that is a link group shared library, triggers a relink, even if the swift file is not part of the link.

**Solution**: We need a new way to provide these linker args, such that they may be propagated and filtered correctly. Additionally, in order to keep the args from exploding in size, we need to utilize a tset and dedupe entries.

To achieve this, we need to
1. pass the library's swift transitive swift module paths if the library compiled swift
2.
    a. in the case of a static library, add it to the static_linkables, which are included with the library's link_info
    b. in the case of a shared library, whether standard or link group, get the extra args from the link_infos being linked

This ensures that the correct swift module files are always included.

The implementation as is, is not ideal. A pattern is clearly forming between how frameworks, swift modules and swift runtimes need to be propagated. As follow ups, I intend to see how we can restructure and generalize this implementation so that its easier to extend and less duplicative.

Reviewed By: milend

Differential Revision: D45366919

fbshipit-source-id: 5deead794870aef6a0f661e21b89ed9a0ab52eb2