Generalize the link_info extract methods in apple_frameworks
Summary:
This method in apple_frameworks.bzl handles frameworks and the swift_runtime. Later in the stack, it will also support the swiftmodules, so in preparation, rename the method.

The structure of these methods and where they exist currently aren't ideal. I'm going to defer on refactoring until its more clear what structure we want.

Reviewed By: milend

Differential Revision: D45379104

fbshipit-source-id: 9a96f076d02422da1948fbe6b05298ec7bc8f17d