Distinguish between project and non-project targets
Summary:
A project, defined via the `haskell_ide` rule, is a set of targets that we want to load at once in the IDE. For example:
```
haskell_ide(
    name = "foo",
    include_projects = [
        ":exe",
        ":lib",
        ":test",
        # dependencies I want to load from source too
        "//common/hs/util:optparse",
    ],
)
```
As seen above, a user may want to load extra targets on the IDE when "opening" their project. But this confuses the IDE's ability to determine which project to load when a source file is open, as there may be multiple project options. This is because the IDE only knows which file is being open, and has to map that to a project

Let's add:

* a new `include_projects` attribute to the `haskell_ide` rule to store the project targets. This will be used by the bxl script to determine project membership.
* a new attribute `extra_projects` to the `haskell_ide` porcelain to allow adding extra targets to the project. These get included in the project targets, but are not used to determine membership.

Then the example above can be written as:
```
haskell_ide(
    name = "foo",
    include_projects = [
        ":exe",
        ":lib",
        ":test",
    ],
    extra_projects = [
        "//common/hs/util:optparse",
    ],
)
```

Reviewed By: donsbot

Differential Revision: D45307588

fbshipit-source-id: 60f68ff4221517d42ee2df84a820e3d89beeeb0e