Use JarBuilderMain when creating a fat_jar
Summary:
If we have native libs, then `fat_jar` works in two stages:
(1) create the `inner.jar`
(2) package `inner.jar`, `FatJarMain.class`, and any native libs into a final `jar`.

Right now, we are using the `jar` tool to accomplish that. This is bad for a couple of reasons:
- The `jar` tool produces a non-deterministic binary, as it doesn't scrub timestamps.
- It compresses the native libs, which is slow.

Instead, we use the existing `jar_builder_main` tool to do the packaging, and we don't compress the native libs.

Reviewed By: ianlevesque

Differential Revision: D44666553

fbshipit-source-id: e6604e72da35dbc6e2572adecfba93b6576e2d7a