Resign the selective debug scrubbed binary if fast adhoc signing was used
Summary:
If the binary is codesigned with fast adhoc signing, we need to resign the executables in order for them to be runnable (else we hit codesigning page faults).

Thus, when fast adhoc signing is enabled, pass a codesign tool for the scrubber to use. In the previous diff, a codesign tool is used, we resign it.

Reviewed By: milend

Differential Revision: D45497126

fbshipit-source-id: 3905f2fb62f9fc633cd6e657af4434f86b8c64e2