Don't add -fdebug-prefix-map when not symlinking headers
Summary:
This was added in D43707948 to provide proper paths in debug info when
using header symlinks. Just using a header map by itself does not use
symlinks though, and the header map will point to the original source
files in that case, so we don't need to do any remapping. Attempting to
perform the remapping in that case also fails because the headers are
source artifacts and so their `owner` property is None, causing a crash
when `_get_debug_prefix_args` attempts to access `path.owner.cell`.

Reviewed By: IanChilds, christycylee

Differential Revision: D45581321

fbshipit-source-id: d06f7a784fd9b374270a89351c01eb589de13d71