Add get_link_execution_preference method
Summary:
The existing `cxx_attr_link_execution_preference` method has two faults:
1. it assumes that `link_execution_preference` is an attribute on the rule on which it is being called. If it's not, we will error with an obscure error:
```
error: Missing parameter `link_execution_preference` for call to ...
```
To alleviate this, let's check if the rule has the attr, and if not, error with a useful error message. This acts as a good safety check that only rules with this attribute attempt to use it.

2. We don't handle the case where `link_execution_preference` is a dependency (always assuming its an enum). So add a `get_link_execution_preference` method to getting the actual enum value. We add support for handling dependencies further up the stack, so until then, let's fail with an unsupported message.

Reviewed By: blackm00n

Differential Revision: D45746135

fbshipit-source-id: 3bba93d8a05ff5ac8c3855da5dcb6b25e6a33036