Compute the link execution preference when linking
Summary:
When linking, we can now determine the link execution preference based upon what is being linked. Make that determination and provide the preference downstream to the linking methods (previous diff).

Unfortunately, this is only available to link group based links, as MergedLinkInfo does not hold onto information on what is part of the link (added TODO).

Now that we know what is being linked, we can:
- Use the link execution preference (if a provided enum)
- If the link execution preference is a dependency, use the `preference_for_links` function to determine the link preference

Right now, no targets provide a function (yet, this is done further up the stack).

Reviewed By: krallin

Differential Revision: D45673509

fbshipit-source-id: 10dfa23d9bbb8cb70e5d0a179868af9e219be4c2