Return extra_linker_outputs as part of the output record
Summary:
When we added `extra_linker_outputs` in D45640114, we returned it as a tuple. It's only used to add them to the propagated subtargets, so abstract that out.

Instead, we can and should return it as part of `_CxxAllLibraryOutputs` (as we will also do with `link_execution_preference_info` in the next diff).

Reviewed By: rmaz

Differential Revision: D45788689

fbshipit-source-id: 48e1b90a77c09a1db6424f6c24a2f4abed54ca09