Refactor build info to use constraints/selects
Summary:
Add constraints to model "stable" v. "full" build info modes used
in fbcode, and refactor the platform definitions and macros to use
`select()`s based on those instead of global config.

This unblocks the subsequent diff to drop build info from the exec
platform, which is the ideal replacement for the current core tools
hack we use.

Reviewed By: luciang

Differential Revision: D45639064

fbshipit-source-id: 91557f070ce97dfb5b639f5282e788b6a909ae30