Expose assets sub-target in js_bundle and js_bundle_genrule
Summary:
There is some built-in behavior in js_bundle and js_bundle_genrule in BUCK to handle the registration of assets when producing an output for Android or iOS. However, we may not want to special case asset handling for all platforms (e.g., Windows).

In order to better support handling of assets on platforms like Windows, this change introduces a sub-target for the assets output of the js_bundle rule.

Reviewed By: IanChilds

Differential Revision: D45974878

fbshipit-source-id: c3f0b5e52b0afde98133b2acbcdf633de85689d8