If we linked locally, scrub locally
Summary:
Currently, the action preference for scrubbing is default. Which means that after linking locally, it will attempt to run on RE, resulting in an upload of the binary to RE.

Instead, if we linked locally, scrub locally. We have the link execution preference, so we can use the same to set whether to scrub locally or not.

Reviewed By: patters

Differential Revision: D46003735

fbshipit-source-id: cd5db6853875fc1026a8e48f5457fa59e6e8e490