Fix threading of Config when using hooks
Summary:
# Problem
The support for hooks was not handling the threading of the Config argument correctly. In particular, the return value of all pre_* functions was being discarded, instead of being passed around to the next callback.

# This diff
The fix we implement is, essentially, to assume that if an `init_per_*` or `pre_{init,end}_per*` callback returns a list, then it is a Config, and pass it along. This covers the weird corner cases, like `ok` being returned when some `pre_*` callbacks are not defined, etc.

Reviewed By: TheGeorge

Differential Revision: D46048776

fbshipit-source-id: 6688b5f92eb9ecadb64345b33d5efca3caf4813d