Support minimize_primary_dex_size without relying on D8
Summary:
Well, this is a little bit of a mess, but we are where we are. For context, read the discussion here: https://fb.workplace.com/groups/redex.feedback/posts/6086680954734971/

Originally, I wanted to rely on `d8` to create the primary dex and the secondary dexes based on the `--main-dex-list`. However, that doesn't work, because it doesn't put the synthetic classes into the primary dex.

We can't use `--main-dex-rules` because that drags in all transitive classes, and our primary dex is not well-defined so we end up trying to put too much in the primary dex.

The solution is to just pass the `.class` files that belong in the primary dex. We do that by splitting every `.jar` into its primary dex `.class` files and the rest. Then we run d8 on the primary dex `.jar`s, and the output is the primary `.dex`.

When we do that, we then need to run `d8` to create the secondary dexes - this means passing the primary dex files as the classpath (to resolve desugaring etc correctly), and it also means passing a secondary dex canary class in the `--main-dex-list`, otherwise `d8` tries to put them all in one dex.

Reviewed By: ianlevesque

Differential Revision: D46072997

fbshipit-source-id: 2702eeeffe4d6863540a0e6d589f0cf69c4cc828