Fix default return value of on_tc_{fail, skip}
Summary:
# Context
The `call_if_exists_with_fallback_store_state/4` function is used for calling both the `pre_*` and `post_*` hook callbacks, but also the `on_tc_fail` and `on_tc_skip` callbacks. The latter two return just a new State, while all the other return a "result" and a new state. To make it work, some special casing is needed for those two.

# Problem
When running a suite that included tests that failed, the state of the CT hooks would get corrupted and the hooks would crash. This was due to the following:

- We would always assume that the default value of all missing functions, including the `on_tc_{fail, skip}` to be a pair with default-result and current-state.
- But we would also then insert a bogus ok-result for `on_tc_{fail, skip}` functions to account for the fact that real ones return no result

So, if, say, calling a missing `on_tc_fail`, the default value returned would be `{ok, State}`, and then we'd additionally wrap it in `{ok, {ok, State}}`, and the state saved would become `{ok, State}`.

# The fix
For the default return result of a missing on_tc_*, we use just the current State, so it returns the same thing like a an existing one would return

Reviewed By: TheGeorge

Differential Revision: D46078362

fbshipit-source-id: 9de58edd5054066a6a54a8f6dd7fed77fa2ac49d