Parse `root` field as `label` instead of `dependency`
Summary:
Previously, the `root` field in the mappings entries passed to
`link_group_map` was always parsed as a `dependency`.  In flows
where we setup potentially large, global mapping files, this means
we eagerly add unnecessary deps to all users of the map file, even
if the mapping would otherwise be irrelevant.

This diff changes to parse the `root` field as `label` to avoid this,
which should prevent unnecessary parse and analysis overhead
for transitive deps of irrelevant roots (and avoids things like eager
visibility or configuration errors).

#forceTDHashing

Reviewed By: asp2insp

Differential Revision: D46022938

fbshipit-source-id: 92717add16b7a2ede6d67a76579c124242cf44cb