Move node count filtering to `LinkGroupInfo` construction
Summary:
This refactors to move `Group` node count filter to before
`LinkGroupInfo` construction, to avoid issues where information
about the filtered links hang around in the preferred linkage dict.
This also avoids needing to maintain a list of disable link groups
to re-attach them to the main link group (which I think this may be
incorrect, as they could of been computed to belong to a non-main
link group in `compute_mappings` if the disable link group had
been removed first).

Reviewed By: christycylee

Differential Revision: D46043251

fbshipit-source-id: 33b8b60847b86312d4fbe78134cdf56635475e39